Time: 17 ms (48.47%), Space: 43.4 MB (72.02%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.java
@@ -25,7 +25,7 @@ class Solution {
         return answer;
     }
 
-    public void DFS(TreeNode node, int sum, int targetSum) {
+    public void DFS(TreeNode node, long sum, int targetSum) {
         if(node == null) {
             return;
         }
